requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altgraph==0.17.4
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
elevenlabslib==0.5.2
frozenlist==1.4.1
idna==3.6
macholib==1.16.3
multidict==6.0.5
numpy==1.26.4
openai==0.27.4
packaging==23.2
py3-tts==3.5
PyAudio==0.2.13
pycparser==2.21
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2024.1
pyobjc==9.0.1
pyobjc-core==9.0.1
pyttsx3==2.90
PyYAML==6.0
requests==2.31.0
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.10.0
tqdm==4.66.2
typing==3.7.4.3
urllib3==2.2.1
yarl==1.9.4
translate
